Create networks from npz, add local multi jobs conversion
